// assets/Models/WalletType.ts
export enum WalletType {
    Particle = "Particle",
    EvmPrivateKey = "EvmPrivateKey",
    SolanaPrivateKey = "SolanaPrivateKey",
    MetaMask = "MetaMask",
    Rainbow = "Rainbow",
    Trust = "Trust",
    ImToken = "ImToken",
    BitKeep = "BitKeep",
    WalletConnect = "WalletConnect",
    Phantom = "Phantom",
    Zerion = "Zerion",
    Math = "Math",
    Omni = "Omni",
    Zengo = "Zengo",
    Alpha = "Alpha",
    Bitpie = "Bitpie",
    Inch1 = "Inch1",
    TokenPocket = "TokenPocket" // iOS not support
}